Refactor scarecrow model to new renderer layout
